requirements.txt
fastapi
uvicorn
motor==3.3.2
beanie==1.25.0
pymongo<4.9.0
pydantic
pydantic-settings
python-dotenv
openai
PyMuPDF
passlib==1.7.4
bcrypt==3.2.2
python-jose[cryptography]
python-multipart
httpx
email-validator
yt-dlp
openai-whisper
